removed inits from test folder
