+++ requirements.txt
@@ -1,2 +1,1 @@
-pytest
 more-itertools
