+++ src/rps/sequence_problems/transitions_and_transversions.py
@@ -8,7 +8,7 @@ Return: The transition/transversion ratio R(s1,s2).
 
 from typing import List
 
-from rps.io_manager import parse_fasta
+from parsing import parse_fasta
 
 
 def calculate_tt_ratio(fasta_data: List[str]) -> float:
